spice-gtk: Enable adding alternative video decoders
The video decoder no longer stores its internal state in the
display_stream struct, or depend on it to know which frame to decode or
store the decoded frame. This way adding alternative implementations
will not pile all their implementation details in display_stream and
they will have more flexibility for their implementation.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>